+++ yala/linters.py
@@ -114,10 +114,11 @@ class Pylint(Linter):
 
     def parse(self, lines):
         """Get :class:`base.Result` parameters using regex."""
-        pattern = re.compile(r"""^(?P<path>.+?)
-                                 :(?P<msg>.+)
-                                 :(?P<line_nr>\d+?)
-                                 :(?P<col>\d+?)$""", re.VERBOSE)
+        pattern = re.compile(r'''
+                             .*?^(?P<path>[^\n]+?)
+                             :(?P<msg>.+)
+                             :(?P<line_nr>\d+?)
+                             :(?P<col>\d+?)$''', re.X | re.M | re.S)
         return self._parse_by_pattern(lines, pattern)